+++ packages/better-auth/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vexcms/better-auth
 
+## 0.1.0-alpha.6
+
 ## 0.1.0-alpha.5
 
 ## 0.1.0-alpha.4